Do not return unexpected nil after a delayed flush
SynCache#[] can return nil instead of entry value if the entry is
expired. A race between two fetch_or_add calls from two RemoteCache
clients and a flush call from a third client that gets converted to
flush_at due to flush_delay could cause RemoteCache#fetch_or_add to
propagate that nil to an unexpecting caller.

Semantic of fetch_or_add implies that it eventually returns a usable
value calculated by either this or another client. Between inefficiency
of re-evaluating the same nil value twice and inconsistency of returning
nil for a key that would yield a non-nil value, we should err on the
side of consistency.